final project cleanup: moved some files out of lib directory, renamed some folders
