+++ src/commands/compression/compress-command.js
