+++ src/commands/compression/decompress-command.js
