+++ src/commands/file-operations/add-command.js
