+++ src/commands/file-operations/cat-command.js
