+++ src/commands/file-operations/cp-command.js
@@ -1,7 +1,7 @@
 
 import path from 'node:path';
 import InvalidInputError from '../../custom-errors/invalid-input-error.js';
-import copyFile from '../../fs-utils/copy-file.js';
+import copyFile from '../../utils/fs-utils/copy-file.js';
 
 const cpCommand = {
     commandName: 'cp',
