+++ src/commands/file-operations/index.js
