+++ src/commands/hash-calculation/hash-command.js
