+++ src/commands/hash-calculation/index.js
