+++ src/commands/os-info/index.js
