+++ src/commands/os-info/os-command.js
