+++ src/utils/fs-utils/check-file-or-folder-exists.js
